requirements.txt
appdirs==1.4.4
astroid==2.5.1
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
distlib==0.3.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
importlib-metadata==3.7.3
install==1.3.4
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.20.1
pandas==1.2.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pipenv==2020.11.15
prompt-toolkit==3.0.17
pycparser==2.20
Pygments==2.8.1
pylint==2.7.2
pylint-flask-sqlalchemy==0.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.4.1
zope.interface==5.3.0
